Add files via upload
The player model has a basic face and body since this won't be viewed at this point. The camera is attached to the arm-with-gun model. That way, the arm-with-gun model need only be moved and the camera should be moved with it? Let me know if this works or if it needs to be changed